requirements.txt
appdirs==1.4.3
boto==2.47.0
certifi==2017.4.17
chardet==3.0.4
configparser==3.5.0
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.10.7
django-allauth==0.32.0
django-braces==1.11.0
django-ckeditor==5.2.2
django-contrib-comments==1.8.0
django-countries==4.6.1
django-embed-video==1.1.2
django-likes==1.11
django-model-utils==3.0.0
django-registration==2.2
django-secretballot==0.7.0
django-star-ratings==0.5.6
django-stdimage==2.4.1
django-storages==1.6.3
django-threadedcomments==1.1
django-tinymce==2.6.0
django-widget-tweaks==1.4.1
idna==2.5
oauthlib==2.0.2
olefile==0.44
packaging==16.8
Pillow==4.1.1
progressbar2==3.30.2
psycopg2==2.7.1
pyparsing==2.2.0
python-utils==2.1.0
python3-openid==3.1.0
redis==2.10.5
requests==2.17.3
requests-oauthlib==0.8.0
six==1.10.0
sorl-thumbnail==12.3
urllib3==1.21.1
whitenoise==3.3.0
